backend/requirements.txt
fastapi>=0.110
uvicorn>=0.29
pydantic>=2.6
networkx>=3.2
chromadb>=0.4.24
google-genai>=1.0
httpx>=0.27
python-multipart
python-dotenv
pytest>=8.0
